[ADD] Roles & Permission to user

+++ app/Http/Controllers/Api/PermissionController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Api;
 use App\Http\Controllers\Controller;
 use App\Http\Resources\PermissionResource;
 use App\Http\Requests\Api\PermissionStoreRequest;
+use App\Http\Requests\PermissionUpdateRequest;
 use App\Models\Permission;
 use Illuminate\Http\Response;
 use Spatie\QueryBuilder\QueryBuilder;
@@ -21,6 +22,7 @@ class PermissionController extends Controller
 
     public function index()
     {
+        abort_if(!user()->tokenCan('permissions_access'), 403);
         $roles = QueryBuilder::for(Permission::class)
             ->allowedFilters(['name'])
             ->allowedSorts(['id', 'name', 'created_at'])
@@ -32,17 +34,18 @@ class PermissionController extends Controller
 
     public function show(Permission $permission)
     {
+        abort_if(!user()->tokenCan('permission_view'), 403);
+
         return new PermissionResource($permission);
     }
 
     public function store(PermissionStoreRequest $request)
     {
-        $role = Permission::create($request->validated());
-        $role->syncPermissions($request->permissions ?? []);
-        return new PermissionResource($role);
+        $permission = Permission::create($request->validated());
+        return new PermissionResource($permission);
     }
 
-    public function update(Permission $permission, PermissionStoreRequest $request)
+    public function update(Permission $permission, PermissionUpdateRequest $request)
     {
         $permission->update($request->validated());
         return (new PermissionResource($permission))->response()->setStatusCode(Response::HTTP_ACCEPTED);
@@ -50,6 +53,7 @@ class PermissionController extends Controller
 
     public function destroy(Permission $permission)
     {
+        abort_if(!user()->tokenCan('permission_delete'), 403);
         $permission->delete();
         return $this->deletedResponse();
     }
