+++ app/Http/Controllers/Api/ProductUnitController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Api\ProductUnitStoreRequest;
+use App\Http\Requests\Api\ProductUnitUpdateRequest;
 use App\Http\Resources\ProductUnitResource;
 use App\Models\ProductUnit;
 use Illuminate\Http\Response;
@@ -13,6 +14,7 @@ class ProductUnitController extends Controller
 {
     public function index()
     {
+        abort_if(!user()->tokenCan('product_units_access'), 403);
         $productUnits = QueryBuilder::for(ProductUnit::with('product'))
             ->allowedFilters(['product_id', 'name', 'price', 'description'])
             ->allowedSorts(['id', 'product_id', 'name', 'price', 'created_at'])
@@ -23,6 +25,7 @@ class ProductUnitController extends Controller
 
     public function show(ProductUnit $productUnit)
     {
+        abort_if(!user()->tokenCan('product_unit_view'), 403);
         return new ProductUnitResource($productUnit);
     }
 
@@ -33,7 +36,7 @@ class ProductUnitController extends Controller
         return new ProductUnitResource($productUnit);
     }
 
-    public function update(ProductUnit $productUnit, ProductUnitStoreRequest $request)
+    public function update(ProductUnit $productUnit, ProductUnitUpdateRequest $request)
     {
         $productUnit->update($request->validated());
 
@@ -42,6 +45,7 @@ class ProductUnitController extends Controller
 
     public function destroy(ProductUnit $productUnit)
     {
+        abort_if(!user()->tokenCan('product_unit_delete'), 403);
         $productUnit->delete();
         return $this->deletedResponse();
     }
