+++ app/Http/Controllers/Api/RoleController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\Api\RoleStoreRequest;
 use App\Http\Requests\Api\RoleUpdateRequest;
 use App\Http\Resources\RoleResource;
+use Illuminate\Support\Facades\Gate;
 use Spatie\Permission\Models\Role;
 use Spatie\QueryBuilder\QueryBuilder;
 
@@ -31,6 +32,7 @@ class RoleController extends Controller
      */
     public function index()
     {
+        abort_if(!user()->tokenCan('roles_access'), 403);
         $roles = QueryBuilder::for(Role::class)
             ->with('permissions')
             ->allowedFilters(['name'])
@@ -49,8 +51,11 @@ class RoleController extends Controller
      */
     public function store(RoleStoreRequest $request)
     {
-        $role = Role::create(['name' => $request->name]);
-        $role->syncPermissions($request->permissions ?? []);
+        $role = new Role();
+        $role->name = $request->name;
+        $role->guard_name = 'web';
+        $role->save();
+        $role->syncPermissions($request->permission_ids ?? []);
 
         return new RoleResource($role);
     }
@@ -62,6 +67,7 @@ class RoleController extends Controller
      */
     public function show(Role $role)
     {
+        abort_if(!user()->tokenCan('role_view'), 403);
         return new RoleResource($role->load('permissions'));
     }
 
@@ -76,7 +82,6 @@ class RoleController extends Controller
     {
         $role->name = $request->input('name');
         $role->save();
-
         $role->syncPermissions($request->input('permissions'));
 
         return new RoleResource($role);
@@ -89,6 +94,7 @@ class RoleController extends Controller
      */
     public function destroy(Role $role)
     {
+        abort_if(!user()->tokenCan('role_delete'), 403);
         $role->delete();
         return $this->deletedResponse();
     }
