+++ app/Http/Controllers/Api/WarehouseController.php
@@ -14,6 +14,7 @@ class WarehouseController extends Controller
 {
     public function index()
     {
+        abort_if(!user()->tokenCan('warehouses_access'), 403);
         $warehouses = QueryBuilder::for(Warehouse::class)
             ->allowedFilters(['name'])
             ->allowedSorts(['id', 'name', 'created_at'])
@@ -24,6 +25,7 @@ class WarehouseController extends Controller
 
     public function show(Warehouse $warehouse)
     {
+        abort_if(!user()->tokenCan('warehouse_view'), 403);
         return new WarehouseResource($warehouse);
     }
 
@@ -43,6 +45,7 @@ class WarehouseController extends Controller
 
     public function destroy(Warehouse $warehouse)
     {
+        abort_if(!user()->tokenCan('warehouse_delete'), 403);
         $warehouse->delete();
         return $this->deletedResponse();
     }
